requirements.txt
python-magic==0.4.15
scikit-allel==1.1.10
numpy==1.15.3
h5py==2.8.0
